+++ package.json
@@ -10,8 +10,12 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.18",
+    "@react-navigation/native-stack": "^6.11.0",
     "react": "18.3.1",
     "react-native": "0.75.2",
+    "react-native-safe-area-context": "^4.10.9",
+    "react-native-screens": "^3.34.0",
     "react-native-snap-carousel": "4.0.0-beta.6",
     "react-native-splash-screen": "^3.3.0"
   },
