+++ src/screens/LoginScreen.tsx
@@ -1,0 +1,16 @@
+import { Text, View } from "react-native";
+
+
+
+
+const LoginScreen = () => {
+    return(
+        <View>
+            <Text>
+                Login Screen
+            </Text>
+        </View>
+    );
+}
+
+export default LoginScreen;
